chore: refactor all files to typescript
