+++ client/package.json
@@ -6,6 +6,9 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@types/jest": "^27.4.0",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
     "axios": "^0.21.4",
     "cloudinary-react": "^1.7.0",
     "dotenv": "^8.2.0",
@@ -15,6 +18,7 @@
     "react-router-dom": "5.2",
     "react-scripts": "4.0.3",
     "sass": "^1.43.4",
+    "typescript": "^4.5.5",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
