+++ client/src/components/containers/Item.tsx
@@ -1,8 +1,8 @@
 import { useContext } from "react";
 import ImageSmall from "../presentational/image-small.component";
 import { useHistory } from "react-router-dom";
-import { ItemsContext, UserContext } from "../../helpers.js/context";
-import Api from "../../services/api.service";
+import { ItemsContext, UserContext } from "../../helpers/context";
+import * as Api from "../../services/ApiService";
 
 const Item = (props) => {
   const { currentUser } = useContext(UserContext);
