+++ client/src/components/form/ImageForm.tsx
@@ -1,6 +1,6 @@
 import { useContext, useState } from "react";
 import { useHistory } from "react-router-dom";
-import { ItemContext } from "../../helpers.js/context";
+import { ItemContext } from "../../helpers/context";
 import ImageLarge from "../presentational/image-lrg.component";
 import Placeholder from "../../assets/add-img.png";
