+++ client/src/components/form/RegisterForm.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useContext } from "react";
-import { LoginContext, UserContext } from "../../helpers.js/context";
-import Api from "../../services/api.service";
+import { LoginContext, UserContext } from "../../helpers/context";
+import Api from "../../services/ApiService";
 
 const RegisterForm = () => {
   const emptyUser = { firstName: "", lastName: "", email: "", password: "" };
