+++ client/src/components/navigational/Navbar.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import CreateIcon from "../presentational/icon-create.component";
+import CreateIcon from "../presentational/IconCreate";
 import HomeIcon from "../presentational/icon-home.component";
 import ProfileIcon from "../presentational/icon-profile.component";
