+++ client/src/components/presentational/ImageLarge.tsx
