+++ client/src/components/presentational/ImageSmall.tsx
