+++ client/src/components/presentational/Logo.tsx
