+++ client/src/pages/CreateItemPage.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { Redirect } from "react-router-dom";
 import ItemSubmitForm from "../components/form/item-submit.form";
-import { LoginContext } from "../helpers.js/context";
+import { LoginContext } from "../helpers/context";
 
 const CreateItem = () => {
   const { isAuthorised } = useContext(LoginContext);
