+++ client/src/pages/LoginPage.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import { Link, Redirect } from "react-router-dom";
 import Logo from "../components/presentational/logo.component";
 import LoginFormComponent from "../components/form/login.form";
-import { LoginContext } from "../helpers.js/context";
+import { LoginContext } from "../helpers/context";
 
 const LoginScreen = () => {
   const { isAuthorised } = useContext(LoginContext);
