+++ client/src/pages/RegisterPage.tsx
@@ -2,7 +2,7 @@ import { Redirect, Link } from "react-router-dom";
 import React, { useContext } from "react";
 
 import RegisterFormComponent from "../components/form/register.form";
-import { LoginContext } from "../helpers.js/context";
+import { LoginContext } from "../helpers/context";
 
 const RegisterScreen = () => {
   const { isAuthorised } = useContext(LoginContext);
